Extract render with mock provider test functionality into utils.tsx
